backend/requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
flask-cors==4.0.0
bcrypt==4.1.2
pyjwt==2.8.0
flask-jwt-extended==4.6.0
requests==2.31.0
python-magic==0.4.27
apscheduler==3.10.4 